requirements.txt
fastapi
uvicorn
openai
pandas
sqlalchemy
kafka-python
apscheduler
python-dotenv